Add safety checks
