+++ src/app/backend/auth/api/common.go
@@ -1,0 +1,33 @@
+// Copyright 2017 The Kubernetes Dashboard Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package api
+
+// ToAuthenticationModes transforms array of authentication mode strings to valid AuthenticationModes type.
+func ToAuthenticationModes(modes []string) AuthenticationModes {
+	result := map[AuthenticationMode]bool{}
+	modesMap := map[string]bool{}
+
+	for _, mode := range []AuthenticationMode{Token, Basic} {
+		modesMap[string(mode)] = true
+	}
+
+	for _, mode := range modes {
+		if _, exists := modesMap[mode]; exists {
+			result[AuthenticationMode(mode)] = true
+		}
+	}
+
+	return result
+}
